Exception handling for service class

+++ src/main/java/com/chirag/basiccrud/service/impl/CloudVendorServiceImpl.java
@@ -1,5 +1,6 @@
 package com.chirag.basiccrud.service.impl;
 
+import com.chirag.basiccrud.exception.CloudVendorNotFoundException;
 import com.chirag.basiccrud.model.CloudVendor;
 import com.chirag.basiccrud.repository.CloudVendorRepository;
 import com.chirag.basiccrud.service.CloudVendorService;
@@ -26,18 +27,24 @@ public class CloudVendorServiceImpl implements CloudVendorService {
 
     @Override
     public String updateCloudVendor(CloudVendor cloudVendor) {
+        if(cloudVendorRepository.findById(cloudVendor.getVendorId()).isEmpty())
+            throw new CloudVendorNotFoundException("Requested cloud vendor does not exist");
         cloudVendorRepository.save(cloudVendor);
         return "Success";
     }
 
     @Override
     public String deleteCloudVendor(String vendorId) {
+        if(cloudVendorRepository.findById(vendorId).isEmpty())
+            throw new CloudVendorNotFoundException("Requested cloud vendor does not exist");
         cloudVendorRepository.deleteById(vendorId);
         return "Success";
     }
 
     @Override
     public CloudVendor getCloudVendor(String vendorId) {
+        if(cloudVendorRepository.findById(vendorId).isEmpty())
+            throw new CloudVendorNotFoundException("Requested cloud vendor does not exist");
         return cloudVendorRepository.findById(vendorId).get();
     }
